tidied up BW impl
